Move callouts to Knowledge Base. (#3902)
* Move callouts to Knowledge Base.

* Set group value on callout creation.

* Remove unused component.
