+++ src/core/i18n/en/translation.en.json
@@ -430,7 +430,8 @@
       "CommunityLeft": "Community Left",
       "CommunityRight": "Community Right",
       "ChallengesLeft": "Challenges Left",
-      "ChallengesRight": "Challenges Right"
+      "ChallengesRight": "Challenges Right",
+      "KnowledgeBase": "Knowledge Base"
     }
   },
   "journey-creation": {
